FE-Fix: SettingsModal 모듈화
