+++ apps/web/components/SettingsModal/index.tsx
@@ -1,0 +1,4 @@
+export { default as SettingsModal } from "./SettingsModal";
+export { default as SettingsModalHeader } from "./SettingsModalHeader";
+export { default as SettingsModalContent } from "./SettingsModalContent";
+export { default as SettingsModalButton } from "./SettingsModalButton";
